Add 'release' interface/script support.
A release is related to bumping the package versions, generating
a DVCS tag, and pushing the changes to remote.

The release script uses https://github.com/peritus/bumpversion and
stores the canonical value in the VERSION file.

In the future, the bumpversion config should be updated to use a
named version scheme to allow for tagging/branching for stages,
e.g. beta/gamma or developer-specific ones for more complex
deployment environments.