Add 'lint' interface/script support.
This target is for running one or more linters against source code
found within the repository.

The default here uses Hadolint and Shellcheck to validate the
Dockerfile and scripts, respectively.

Note: Dockerfile presence is optional (not all repos will use Docker) so
either continue to make it optional based on the file or figure out
the best way to make this conditional within Cookiecutter itself.